+++ .env
@@ -1,2 +1,3 @@
+PORT=4000
 SALT=10
 SECRET_SIGNATURE=Its a secret
