+++ package.json
@@ -4,7 +4,8 @@
     "private": true,
     "main": "dist/index.js",
     "scripts": {
-        "start": "ts-node-dev src/index.ts",
+        "start": "ts-node src/index",
+        "dev": "ts-node-dev src/index.ts",
         "test": "jest --watch"
     },
     "dependencies": {
@@ -30,14 +31,15 @@
         "@types/helmet": "^0.0.45",
         "@types/jest": "^24.0.24",
         "@types/jsonwebtoken": "^8.3.5",
-        "@types/node": "^12.12.7",
+        "@types/node": "^13.1.4",
         "@types/pg": "^7.11.2",
         "@types/sqlite3": "^3.1.5",
         "@types/supertest": "^2.0.8",
         "jest": "^24.9.0",
         "supertest": "^4.0.2",
         "ts-jest": "^24.2.0",
-        "ts-node": "^8.4.1",
+        "ts-loader": "^6.2.1",
+        "ts-node": "^8.5.4",
         "ts-node-dev": "^1.0.0-pre.44",
         "typescript": "^3.7.2"
     }
